social_media_api/requirements.txt
# files required for the social_media_api application
Gunicorn>=20.1.0
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=6.0
Pillow>=9.5.0
psycopg2-binary>=2.9.6
django-cors-headers>=4.0.0
pytest-django>=4.5.2
pytest>=7.4.0
Faker>=18.9.0
factory
